requirements.txt
Flask
gunicorn
openai
PyMuPDF
python-dotenv
reportlab
